# api/swagger.yaml
basePath: /api
definitions:
...
  grocademy_internal_db_models.Module:
    properties:
      course_id:
        description: Foreign key to Course
        type: integer
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      order:
        description: Module order within the course
        type: integer
      pdf_content:
        description: Path to the stored PDF file
        type: string
      title:
        type: string
      updated_at:
        type: string
      video_content:
        description: Path to the stored video file
        type: string
    type: object
...
  internal_api_handlers.ReorderModulesRequest:
    properties:
      module_order:
        items:
          properties:
            id:
              type: integer
            order:
              type: integer
          required:
          - id
          - order
          type: object
        type: array
    required:
    - module_order
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Fachriza Ahmad Setiyono - Seleksi Labpro 3
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Grocademy API
  version: "1.0"
paths:
...
  /courses/{courseId}/modules:
    get:
      description: Retrieve a list of all modules for a given course, with optional
        pagination and search parameters
      parameters:
      - description: Course ID
        in: path
        name: courseId
        required: true
        type: integer
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Items per page (default 10)
        in: query
        name: limit
        type: integer
      - description: Search query
        in: query
        name: q
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/grocademy_internal_db_models.Module'
            type: array
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get all modules for a specific course with pagination and search
      tags:
      - modules
    post:
      consumes:
      - multipart/form-data
      description: Create a new module from multipart form data for the given course
        ID
      parameters:
      - description: Course ID
        in: path
        name: courseId
        required: true
        type: integer
      - description: Module title
        in: formData
        name: title
        required: true
        type: string
      - description: Module description
        in: formData
        name: description
        required: true
        type: string
      - description: Module order within the course
        in: formData
        name: order
        required: true
        type: integer
      - description: PDF file for module content
        in: formData
        name: pdf_content
        type: file
      - description: Video file for module content
        in: formData
        name: video_content
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/grocademy_internal_db_models.Module'
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Course not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Create a new module for a specific course
      tags:
      - modules
  /courses/{courseId}/modules/reorder:
    patch:
      consumes:
      - application/json
      description: Update the order of multiple modules for a specific course
      parameters:
      - description: Course ID
        in: path
        name: courseId
        required: true
        type: integer
      - description: List of module IDs and their new orders
        in: body
        name: module_order
        required: true
        schema:
          $ref: '#/definitions/internal_api_handlers.ReorderModulesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 'message: Modules reordered successfully'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid input or modules not belonging to course
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Reorder modules within a course
      tags:
      - modules
...
  /modules/{id}:
    delete:
      description: Deletes a module record by ID (soft delete)
      parameters:
      - description: Module ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Module deleted successfully
        "400":
          description: Invalid module ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Module not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Delete a module
      tags:
      - modules
    get:
      description: Retrieve a single module by its ID
      parameters:
      - description: Module ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/grocademy_internal_db_models.Module'
        "400":
          description: Invalid module ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Module not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get a module by ID
      tags:
      - modules
    put:
      consumes:
      - multipart/form-data
      description: Update specified fields of a module by ID, with optional file uploads
      parameters:
      - description: Module ID
        in: path
        name: id
        required: true
        type: integer
      - description: Module title
        in: formData
        name: title
        type: string
      - description: Module description
        in: formData
        name: description
        type: string
      - description: Module order within the course
        in: formData
        name: order
        type: integer
      - description: New PDF file for module content
        in: formData
        name: pdf_content
        type: file
      - description: New Video file for module content
        in: formData
        name: video_content
        type: file
      - description: Set to true to clear existing PDF content
        in: formData
        name: clear_pdf
        type: boolean
      - description: Set to true to clear existing Video content
        in: formData
        name: clear_video
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: Updated module object
          schema:
            $ref: '#/definitions/grocademy_internal_db_models.Module'
        "400":
          description: Invalid input or no fields to update
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Module not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Update a module's data
      tags:
      - modules
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"